Reports: map layers and legend
