+++ app/layout/scenarios/reports/solutions/page-3/constants.ts
