--- app/layout/scenarios/reports/solutions/page-4/index.ts
@@ -1,1 +1,0 @@
-export { default } from './component';
