+++ app/pages/reports/[pid]/[sid]/solutions.tsx
@@ -8,8 +8,7 @@ import ReportHeader from 'layout/scenarios/reports/header';
 import Page1 from 'layout/scenarios/reports/solutions/page-1';
 import Page2 from 'layout/scenarios/reports/solutions/page-2';
 import Page3 from 'layout/scenarios/reports/solutions/page-3';
-import Page4 from 'layout/scenarios/reports/solutions/page-4';
-import WebShootStatus from 'layout/scenarios/reports/webshot-status';
+import WebShotStatus from 'layout/scenarios/reports/webshot-status';
 
 export const getServerSideProps = withProtection(withUser());
 
@@ -54,16 +53,7 @@ const SolutionsReport: React.FC = () => {
         <Page3 />
       </div>
 
-      <div
-        style={styles.page}
-        className="flex flex-col h-full text-black bg-white"
-      >
-        <ReportHeader />
-        <Page4 />
-      </div>
-
-      <WebShootStatus />
-
+      <WebShotStatus />
     </>
   );
 };
